Move DfeSignIn::UserAccess to DfeSignIn::Api::User
Better reflects what this class does: RESTful GET action for a User.
