+++ lib/dfe_sign_in.rb
@@ -18,4 +18,4 @@ module DfeSignIn
 end
 
 require_relative "dfe_sign_in/utils"
-require_relative "dfe_sign_in/user_access"
+require_relative "dfe_sign_in/api/user"
